# artifacts/vercel-app/app/api/orders/route.ts
import { createClient } from "@supabase/supabase-js";
import { NextRequest, NextResponse } from "next/server";

function getSupabase() {
  return createClient(
    process.env.NEXT_PUBLIC_SUPABASE_URL!,
    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
  );
}

function toCamelCase(obj: Record<string, unknown>): Record<string, unknown> {
  return Object.fromEntries(
    Object.entries(obj).map(([k, v]) => [
      k.replace(/_([a-z])/g, (_, l) => l.toUpperCase()),
      v,
    ])
  );
}

function transformOrders(orders: Record<string, unknown>[]): Record<string, unknown>[] {
  return orders.map(toCamelCase);
}

export async function GET() {
  const supabase = getSupabase();
  const { data, error } = await supabase.from("orders").select("*").order("created_at", { ascending: false });
  if (error) {
    return NextResponse.json({ error: error.message }, { status: 500 });
  }
  return NextResponse.json(transformOrders(data || []));
}

function toSnakeCase(obj: Record<string, unknown>): Record<string, unknown> {
  return Object.fromEntries(
    Object.entries(obj).map(([k, v]) => [
      k.replace(/([A-Z])/g, "_$1").toLowerCase(),
      v,
    ])
  );
}

export async function POST(request: NextRequest) {
  const supabase = getSupabase();
  try {
    const body = await request.json();
    const snakeBody = toSnakeCase(body);
    const { data, error } = await supabase.from("orders").insert(snakeBody).select().single();
    if (error) {
      return NextResponse.json({ error: error.message, details: error.details, hint: error.hint }, { status: 400 });
    }
    return NextResponse.json(toCamelCase(data), { status: 201 });
  } catch (e) {
    console.error("Order parse error:", e);
    return NextResponse.json({ error: String(e) }, { status: 400 });
  }
}