# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jamuhuri Gachoroba Author & Podcast API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BookType = (typeof BookType)[keyof typeof BookType];

export const BookType = {
  hardcopy: "hardcopy",
  ebook: "ebook",
  both: "both",
} as const;

export interface Book {
  id: number;
  title: string;
  subtitle?: string | null;
  description: string;
  author: string;
  coverImage?: string | null;
  type: BookType;
  hardcopyPrice?: number | null;
  ebookPrice?: number | null;
  currency: string;
  isLatest: boolean;
  publishedYear?: number | null;
  category?: string | null;
  createdAt: string;
}

export type CreateBookBodyType =
  (typeof CreateBookBodyType)[keyof typeof CreateBookBodyType];

export const CreateBookBodyType = {
  hardcopy: "hardcopy",
  ebook: "ebook",
  both: "both",
} as const;

export interface CreateBookBody {
  title: string;
  subtitle?: string | null;
  description: string;
  author: string;
  coverImage?: string | null;
  type: CreateBookBodyType;
  hardcopyPrice?: number | null;
  ebookPrice?: number | null;
  currency: string;
  isLatest: boolean;
  publishedYear?: number | null;
  category?: string | null;
}

export type CreateOrderBodyOrderType =
  (typeof CreateOrderBodyOrderType)[keyof typeof CreateOrderBodyOrderType];

export const CreateOrderBodyOrderType = {
  hardcopy: "hardcopy",
  ebook: "ebook",
} as const;

export interface CreateOrderBody {
  bookId: number;
  bookTitle: string;
  orderType: CreateOrderBodyOrderType;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  deliveryAddress?: string | null;
  deliveryCity?: string | null;
  quantity?: number | null;
  totalAmount?: string | null;
  vatAmount?: string | null;
  notes?: string | null;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
  notes?: string | null;
}

export type OrderOrderType =
  (typeof OrderOrderType)[keyof typeof OrderOrderType];

export const OrderOrderType = {
  hardcopy: "hardcopy",
  ebook: "ebook",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  bookId: number;
  bookTitle: string;
  orderType: OrderOrderType;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  deliveryAddress?: string | null;
  deliveryCity?: string | null;
  status: OrderStatus;
  notes?: string | null;
  createdAt: string;
}

export interface Podcast {
  id: number;
  title: string;
  description: string;
  audioUrl?: string | null;
  buzzsproutUrl?: string | null;
  duration?: string | null;
  episodeNumber?: number | null;
  season?: number | null;
  publishedAt: string;
  thumbnailUrl?: string | null;
  tags: string[];
  createdAt: string;
}

export interface CreateSubscriberBody {
  name: string;
  email: string;
  phone?: string | null;
  wantsWhatsapp: boolean;
}

export interface Subscriber {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  wantsWhatsapp: boolean;
  whatsappApproved: boolean;
  subscribedAt: string;
}

export type CreateMessageBodyType =
  (typeof CreateMessageBodyType)[keyof typeof CreateMessageBodyType];

export const CreateMessageBodyType = {
  order: "order",
  contact: "contact",
  subscription: "subscription",
} as const;

export interface CreateMessageBody {
  type: CreateMessageBodyType;
  subject: string;
  body: string;
  senderEmail?: string | null;
}

export type MessageType = (typeof MessageType)[keyof typeof MessageType];

export const MessageType = {
  order: "order",
  contact: "contact",
  subscription: "subscription",
} as const;

export interface Message {
  id: number;
  type: MessageType;
  subject: string;
  body: string;
  senderEmail?: string | null;
  readAt?: string | null;
  createdAt: string;
}

export interface StatsSummary {
  totalBooks: number;
  totalOrders: number;
  totalSubscribers: number;
  totalPodcasts: number;
}

export type DeleteBook200 = {
  success: boolean;
};

export interface SitePage {
  id: number;
  pageName: string;
  pageTitle: string;
  heroTitle?: string | null;
  heroSubtitle?: string | null;
  heroDescription?: string | null;
  heroImage?: string | null;
  heroButton?: string | null;
  heroButtonText?: string | null;
  bodyContent?: string | null;
  footerContent?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  socialLinks?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface UpdatePageBody {
  pageName: string;
  pageTitle: string;
  heroTitle?: string | null;
  heroSubtitle?: string | null;
  heroDescription?: string | null;
  heroImage?: string | null;
  heroButton?: string | null;
  heroButtonText?: string | null;
  bodyContent?: string | null;
  footerContent?: string | null;
  phone?: string | null;
  email?: string | null;
  address?: string | null;
  socialLinks?: string | null;
}