# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Jamuhuri Gachoroba Author & Podcast API
servers:
  - url: /api
    description: Base API path
...
paths:
  /pages:
    get:
      operationId: listPages
      tags: [pages]
      summary: List all site pages
      responses:
        "200":
          description: List of pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SitePage"
    /pages/{name}:
    get:
      operationId: getPage
      tags: [pages]
      summary: Get a site page by name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Page content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SitePage"
        "404":
          description: Page not found
    put:
      operationId: updatePage
      tags: [pages]
      summary: Update or create a site page (admin)
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePageBody"
      responses:
        "200":
          description: Page updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SitePage"
        "201":
          description: Page created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SitePage"
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new book order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid request

    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
  /subscribers:
    post:
      operationId: createSubscriber
      tags: [subscribers]
      summary: Register a new podcast subscriber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriberBody"
      responses:
        "201":
          description: Subscriber registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscriber"
        "409":
          description: Already subscribed

    get:
      operationId: listSubscribers
      tags: [subscribers]
      summary: List all subscribers (admin)
      responses:
        "200":
          description: Subscriber list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscriber"
...
components:
  schemas:
...
    CreateOrderBody:
      type: object
      properties:
        bookId:
          type: integer
        bookTitle:
          type: string
        orderType:
          type: string
          enum: [hardcopy, ebook]
        customerName:
          type: string
        customerEmail:
          type: string
          format: email
        customerPhone:
          type: string
          nullable: true
        deliveryAddress:
          type: string
          nullable: true
        deliveryCity:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - bookId
        - bookTitle
        - orderType
        - customerName
        - customerEmail
...
    Order:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        bookTitle:
          type: string
        orderType:
          type: string
          enum: [hardcopy, ebook]
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
          nullable: true
        deliveryAddress:
          type: string
          nullable: true
        deliveryCity:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - bookId
        - bookTitle
        - orderType
        - customerName
        - customerEmail
        - status
        - createdAt
...
    CreateSubscriberBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        wantsWhatsapp:
          type: boolean
      required:
        - name
        - email
        - wantsWhatsapp
    Subscriber:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        wantsWhatsapp:
          type: boolean
        whatsappApproved:
          type: boolean
        subscribedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - wantsWhatsapp
        - whatsappApproved
        - subscribedAt
...
    SitePage:
      type: object
      properties:
        id:
          type: integer
        pageName:
          type: string
        pageTitle:
          type: string
        heroTitle:
          type: string
          nullable: true
        heroSubtitle:
          type: string
          nullable: true
        heroDescription:
          type: string
          nullable: true
        heroImage:
          type: string
          nullable: true
        heroButton:
          type: string
          nullable: true
        heroButtonText:
          type: string
          nullable: true
        bodyContent:
          type: string
          nullable: true
        footerContent:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        socialLinks:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - pageName
        - pageTitle
        - createdAt
        - updatedAt
    UpdatePageBody:
      type: object
      properties:
        pageName:
          type: string
        pageTitle:
          type: string
        heroTitle:
          type: string
          nullable: true
        heroSubtitle:
          type: string
          nullable: true
        heroDescription:
          type: string
          nullable: true
        heroImage:
          type: string
          nullable: true
        heroButton:
          type: string
          nullable: true
        heroButtonText:
          type: string
          nullable: true
        bodyContent:
          type: string
          nullable: true
        footerContent:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        socialLinks:
          type: string
          nullable: true
      required:
        - pageName
        - pageTitle